Fix deserialization of LootTableEntry
Because the LootTableEntry fields are readonly, the deserializer seems not to be able to properly create entries. This can be observed by adding a debug log to the Add function: `PrefixLogHandler.Get("[LT]").Debug($"add {item.Item}");` gives log entries like:
```
[Debug  :Subnautica Randomiser] [LT] add None
[Debug  :Subnautica Randomiser] [LT] add None
[Debug  :Subnautica Randomiser] [LT] add None
[Debug  :Subnautica Randomiser] [LT] add None
[Debug  :Subnautica Randomiser] [LT] add None
[Debug  :Subnautica Randomiser] [LT] add None
...
```

After the patch, it instead looks like:
```

[Debug  :Subnautica Randomiser] [LT] add CrashPowder
[Debug  :Subnautica Randomiser] [LT] add Copper
[Debug  :Subnautica Randomiser] [LT] add ScrapMetal
[Debug  :Subnautica Randomiser] [LT] add Quartz
[Debug  :Subnautica Randomiser] [LT] add Salt
[Debug  :Subnautica Randomiser] [LT] add Titanium
...
```

and the behavior works as expected. As to why this changed in 0.12.0, not sure, but I built and tested the fix, and verified that it worked.